feat: Implement asteroid splitting and collision handling
Add functionality for asteroids to split into smaller ones upon being shot.
Introduce collision detection for player and shot interactions with asteroids.
Refactor game state logging and update player shooting mechanics with cooldown.

+++ src/astropy/entities/physics_body.py
@@ -18,7 +18,7 @@ class PhysicsBody(pygame.sprite.Sprite, ABC):
 
         self.position: pygame.Vector2 = pygame.Vector2(x, y)
         self.velocity: pygame.Vector2 = pygame.Vector2(0, 0)
-        self.radius: float = radius
+        self.radius: int = radius
 
     @abstractmethod
     def draw(self, screen: pygame.Surface) -> None:
